requirements.txt
requests>=2.31.0
rich>=13.7.0
langdetect>=1.0.9
fastapi>=0.115.0
uvicorn>=0.30.0
Pillow>=10.0.0
pillow-heif>=1.1.0
pytesseract>=0.3.10
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
PyJWT>=2.10.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
authlib>=1.3.0
itsdangerous>=2.1.0
